requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
lightgbm==4.0.0
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.1
python-binance==1.0.17
requests==2.31.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
sqlite3
tqdm==4.65.0
vaderSentiment==3.3.2